replace globby with tinyglobby
vitest trips up on globby having a default section in package.json exports, but not an mjs file.

There is a TODO in the codebase to handle this situation, but who knows when it will be resolved.

tinyglobby is used by vite and many other projects, so this seems like a safe enough replacement