fix: Partial hack fix for kysely-codegen #110
Right now it's hard to use this module as a dependency for something else since it attempts to pull in imports we don't
have and don't want to have. (@libsql/libsql and bun:sqlite). The fix needs to be done in kysely-codegen itself I
imagine.

In the meantime, just export a separate codegen target. The consumer can then install the dependencies

https://github.com/RobinBlomberg/kysely-codegen/issues/110

If this can't be fixed upstream, we should probably redo things here and
just export function to pass to an already installed kysely-codegen.